examples/electron/package.json: Run electron-rebuild in postinstall
electron-rebuild is currently called in the build:electron target.  This
means that it's called every time we rebuild our code, even though it
just needs to be called once, just after "npm install" (it can be seen
as a fixup to "npm install").  This patch moves the call to the
postinstall step.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>