examples/*/package.json: Remove ./node_modules/.bin prefix
In the npm scripts, we can use foo instead of ./node_modules/.bin/foo,
because the ./node_modules/.bin directory is automatically added to the
PATH.  It helps readability by reducing the length of the lines, which
are already quite long.

The only downside I see is that if you copy paste the line from the npm
output to run it manually, you'll need to add ./node_modules/.bin/ in
from of the program names, or add it to your PATH.  I think the
readability increase is more valuable though.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>