examples/*/package.json: Add "start" targets
Currently, start:browser starts everything in development mode.  I think
it would make more sense if it only started the browser part (including
the development web server) and if the start target was used to start
the whole thing.  "npm start" is a standard top-level npm command which
invokes the "start" script, so people are used to it.

I also replaced ./node_modules/.bin/concurrently with simply
concurrently.  In the npm scripts, we can use foo instead of
./node_modules/.bin/foo, because the ./node_modules/.bin directory is
automatically added to the PATH.  It helps reducing the length of the
lines.

Finally, I think leaving some blank lines between logical groups help
readability a bit.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>