Add debug logging support
This patch adds debug logging support to the frontend and backend.

It adds a logging interface over JSON-RPC to a bunyan based server on the
backend.

This can be accessed by injecting a ILogger into a class and calling
bunyan like functions like so:

this.logger.debug('This is a debug message);
this.logger.info('This is an info message);

Note that since this interface is over JSON-RPC the log functions return a
Promise but since logs can be async, waiting for this promise to resolve
is not needed and it can be called as a normal function call.

Note also that to minimize the traffic over JSON-RPC and string
concatenation the log level is cached by the frontend so that the logs are
not sent over JSON-RPC if they do not meet the cached minimum level.

This cache is reset if functions that modify the minimum level are called.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>