[tests]configure nyc to report coverage on all source files
This is so we get the most accurate test coverage report possible.
Before this, nyc was configured to only report coverage for source
files that are touched by the tests. So a file that is not tested
at all would not be reported, instead of reporting that it's 0%
covered. Also, for those packages that do not have any unit tests,
it was necessary to add a bogus one just to that mocha doesn't
fail the tests.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>