[quality] add strict.tslint.json (fixes #356)
This new tslint configuration file extends dev.tslint.json and
adds many rules that might or might not be a good fit for Theia.
To use this file while developing Theia in vscode, change file
$THEIA/.vscode/settings.json and use it for "tslint.configFile".
Note that this will greatly increase the number of reported
linter warnings, some of which may not make sense.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>