Task extension
This extension provides the means to execute scripts or executables,
from a frontend client, in the backend. Tasks life cycle events are
send from the backend to the frontend clients, whenever a task is
created or terminates. A task client can ask the backend for the list
of tasks that are currently running.

We support having multiple frontend task clients that are using the same
backend task server. Each client uses its workspace root as context
parameter, in all its communications with the backend. Whenever the
backend sends notifications to its clients, the original client context
is sent along. It's then up-to the clients to determine if they are
interested in the event, given its context, and show it to the user or
not.

This mechanism makes it possible for multiple clients, opened on the
same workspace, to all get the notifications related to all tasks that
execute from that workspace. At the same time, clients can ignore tasks
that are run from another workspace.

Fixes #686

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>