Add base process classes
This patch adds classes to wrap around a process created by node-pty or
node's child_process.

These are called respectively TerminalProcess and RawProcess they both
derive from a base class called Process.

This is needed so that we may extend derived FooProcess classes that spawn
processes in a specific way and can expose custom interfaces while keeping
a base interface to manage the processes in a generic way.

An example of derived class  is the ShellProcess to be commited later in this
PR.

This part of the implementation of issue #312

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>