Reduce the number of config builds on CI
This fixes #451, in order to speed up CI we decided to reduce the number
of configs.

Theia will be built on node 8 for linux and osx and on node 6 for linux only.

Node 7 is dropped.

This patch also cleans up the secure key from the matrix, it's not needed
since it's kept global.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>