[terminal] set padding on terminal-container class
In terminal.css we set some padding so that the terminal's text
does not start right at the upper-left corner of the containing
widget. It used to be set on class xterm-row, but it seems that
this had the side-effect of confusing xterm.js' copy listener
into mis-mapping the mouse coordinates to corresponding text to
copy, resulting in the copied text being shifted compared to the
selection. Putting the padding on class terminal-container does
not have the same problem. Fixes #460

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>