Remove sourcing of the .profile and cd root-dir workaround
Before this patch when starting a shell 2 commands were sent directly on
the terminal source .profile and cd root-dir.

This patch removes this and uses cwd to change to the root-dir and just
removes the sourcing of the .profile.

A cleaner way to do the .profile sourcing needs to be found on macOS.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>