Fix terminal initial size
This patch fixes the terminal initialize since, before this patch the
initial geometry calculated by (this.term as any).proposeGeometry() was
undefined causing default values to be used for the terminal size.

This was caused by a call to getComputedStyle on the parent element before
the element was actually in the DOM.

This patch makes it so that the widget is added to the DOM and only after
this is done the terminal is started and its geometry is calculated.

Fixes: #294

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>