Make it possible to export container dependant variables
This patch introduces the initialize method in
Frontend/BackendApplicationContribution this is called as the first thing
the application does before onStart or configure.

This enables the logger module to export a logger object fetched via
container.get<ILogger>(ILogger) such that this new logger object can be
exported to code that does not use injection.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>