Refs #516: Upgrade xterm.js to v3 branch
There were many tmux-related fixes in the v3 branch (the next major
version, they are currently at 2.x.y).  This patch updates the
dependencies of the terminal module such that it uses the latest
unstable/unreleased/development version of xterm.js, branch v3.

We can point the dependency in the package.json directly to a github
repo (that's what theia-ide/xterm.js does).  However, running "yarn"
just checks out the repo, it doesn't build it.  Therefore, I created a
fork of xterm.js in the theia-ide organization, created a new branch
(v3-built), built it (npm install) and checked in the result (the lib/
directory).  We can use that branch as a dependency.

Some minor changes in the terminal module were needed due to API
changes.

An ignore-loader entry was added in the webpack config (generator) to
get rid of a warning about webpack trying to import an html file.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>