Temporarly skip extension-package outdated test
With this test enabled, the build breaks if we publish a version that is
more recent than next.

So each time we publish a new version this will fail.

This needs to be refactored, but this patch skips the tests to allow tests
to pass until this is properly fixed.

See #739 for more information.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>