inotify file watching: Only print ENOSPC error once
When the limit of inotify file watches has been reached, one error
will be printed for each file that we subsequently try to watch.  This
floods the console, which is not very useful.  Instead, print the error
message only once, with a link to our wiki with indications on how to
raise the inotify limit.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>