Refs #378: Very simple adaptive tab size
Currently, all tabs have the same width, regardless of their content.
This means that long filenames are cut short.  While far from ideal, I
managed to make the situation a little bit better by setting flex-basis
(the 3rd parameter of flex) to auto.  The flex attribute now has the
default value, so we could as well remove it, but I think it's good to
leave it here for clarity.

With this patch, if there is enough room, tabs will show the entire
filename.  When there isn't enough space for all tabs, they will start
shrinking.  Prior to this patch, tabs had a fixed size, until there was
no more space, where they also started shrinking.

This is not ideal, because if you open many files, the tabs will be so
small that you won't be able to read any filename.  I still think that
this patch is a small improvement, because it improves the case where
you have few files open, while not changing the case where you have
many.

I also removed the min-width attribute of
.p-TabBar-tab.p-mod-drag-image.  Otherwise, when dragging a narrow tab
(with a small filename), a large drag image would appear.  Without it,
the drag image looks pretty much like the normal tab (but without the
X).

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>