Keep terminal history when restoring a terminal
This patch allows the terminal to keep its history when being restored,
this also works when restoring the same workspace in multiple tabs such
that you can have multiple terminal widgets connected to the same
underlying backend terminal.

Important to note that closing the terminal by closing the tab will close
the backend terminal killing the whole browser tab or refreshing doesn't,
the terminal is then detached and can be later restored.

Fixes: #821

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>