Add keymaps package
This allows the user to rebind commands with custom keybindings in a
file located in the user_storage location i.e
"user_storage:keymaps.json".

It also adds a menu contribution that the user can easily open with
File->Open Keyboard Shortcuts, which in turns calls the user storage
service for the resource.

It also adds easy naming to keybindings and
makes rebinding commands easier by only having to type "ctrl+a" instead
of "MODIFIER.M1+KeyA" or something like that.

Default keybindings are mapped in a structure called KeybindingIndex,
which allows keybindings to be "shadowed" (remapped) by another
KeybindingIndex. This way, eventually users could download an eclipse
keymaps file or an emacs keymaps file which would automatically shadow
the default index (or at least bindings it can shadow).

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@ericsson.com>