Clarify error when language server fails to initialize
When a language server fails to launch, an error message "Connection got
disposed" is shown to the user. This could be confusing because it could mean
that the connection to the backend was lost. This commit changes the message to
"Failed to start language server $NAME".

Further improvements would be to display full instructions to install the server
by adding actions to the notification and also possibly an action to never show
the message again. This could be addressed in separate commits.

Fixes #982

Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>