Fixes #1002
The test creates a KeyCode object by passing a sequence composed of
keycode numbers (platform agnostic) and Modifiers (theia enum). However,
the constructor then checks which platform it's currently running and
creates the appropriate KeyCode object (Modifier.M1 sets the ctrl
property to true on Linux/Windows but the meta property on OS X).
Instead of adding an OSX check everywhere, this changes the test parsing
methods to use shift instead, which is platform agnostic, and adds
specific OSX/Other platform tests to make sure it would work on them.

Also disables the tslint no-unused-expression in the spec file which
makes it very difficult to develop otherwise because of 'expect'.

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@ericsson.com>