Update xterm.js
The main reason for updating xterm.js is to get the fixes for the wrong
height and width computation.  Now, the output should not be too long or
to short for the available space, both vertically and horizontally.

Because of how things are computed now, it is not possible to apply a
padding on the terminal container as we currently do.  The padding will
work, but the canvas width won't be adjusted, so the output will go too
far on the right and at the bottom.  Once this PR

  https://github.com/xtermjs/xterm.js/pull/1123

gets merged in xterm.js, we should be able to add back the padding.

The addon system has been reworked.  Before, just importing the addons
(fit and attach) would modify the Terminal prototype and add some
methods.  Now, we have to call the static method Terminal.applyAddon.

Also, because of the changes to the addon system, we don't need the
ignore-loader snippets in the webpack configuration anymore.

Finally, I think we don't need @types/xterm anymore, since the xterm
package provides its own typings.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>