Make raw-process handle non-executable paths gracefully
When Node's child_process.spawn tries to execute a non-executable path,
it throws an error, rather than calling the error handler registered
with on('error').  We don't handle this currently.

This patch tries to normalize the error handling by catching errors
thrown by child_process.spawn and calling the error handler with it.

I added a test for that.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>