Make process exit logging debug instead of info
The quick search in workspace PR showed that having process exits logged
with the info log level was a bit too verbose.  This patch scales it
down to debug.

It has been suggested to log zero exit code as at the debug log level,
and non-zero exit codes as warnings.  However, a non-zero exit code does
not necessarily mean an error.  For example, with ripgrep, a search that
yields no result returns a non-zero exit code.  Logging this as an error
would be too alarmist, because nothing actually went wrong.  I think it
should be up to the calling code to log this as an error, if they know
that the non-zero exit code represents an error.

In order to make sense out of the log messages though, we need to know
what the process was.  I added the process command and args to the
message.  Passing the arguments from RawProcess and TerminalProcess to
Process hit this:

https://github.com/inversify/InversifyJS/issues/522

I therefore needed to add the @unmanaged() annotations to the new
parameters.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>