Use RawKeybindings to specifiy keybindings
This patch makes extensions use RawKeybindings instead of Keybindings to
specify keybindings.

For the moment the registerKeybinding(s) function is still compatible with
the old Keybinding inputs.  But this will be changed when all extensions
are changed and when serialization is done properly for keybindings.


Also this patch changes the "ControlOrCommand" string to specify the M1
motifier (cmd on OSX and ctrl on win/linux) to "ctrlcmd".

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>