Fixes #1121: Make WorkspaceUriLabelProviderContribution.getLongName return full path if file is not in workspace
When hovering an editor tab for a file that isn't in the workspace, we
currently show nothing.  This is because
WorkspaceUriLabelProviderContribution.getLongName blindly strips the
workspace root from the front of the file path, even if the file isn't
actually in the workspace.  Presumably, if the original file path was
longer than the workspace root, we would be left with a partial path,
but I didn't try it.

This patch makes getLongName check if the original file path starts with
the workspace root before stripping it.  If not, it calls the base class
implementation, which returns the full path.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>