Support keysequences
This patch enables support for key sequences so that for example
keybindings like this one are supported:

"Ctrl-u t", meaning press Ctrl-u and then press t.

Here's a few limitations at the moment however:

  - These keybindings can't be shown as part of an electron MenuItem see:
  https://github.com/electron/electron/issues/11740#issuecomment-360980460

  - Monaco only supports a sequence of 2 keybinding so "Ctrl-u t" will
    work but not "Ctrl-u t a"

    This means that for now in the QuickMenu if the sequence is more then
    2 keychords long we don't show it.

Also note that at the moment the delay to enter a sequence is hardcoded to 1 sec
between keystrokes.

Fixes: #472

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>