Remove timeout for keybindings replace by quit key
Instead of a using a timeout to know if a key sequence is done, just check
if any keysequence is possible, if any is possible keep waiting for user
input unless he presses the cancel key.

Pressing ctrlcmd+q will cancel the keybinding wait since no keybinding can
contain the quit keybinding in it's sequence.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>