[tab-bars] compute tab sizes only when attached also call during onAfterAttach
The computation of tabsizes possibly happens before the shell got attached, resulting in width and height computation to be 0px. So the tabs were not visible.
This change ensures that the computation happens on update but also on after attach, so it no longer can happen. It was not reliably reproduceable (I could with a local configuration) but clear when analyzing the flow of events.

The change to initializeLayout is related in that any async code should return a promise so the caller can decide whether to wait for the execution to happen.

Signed-off-by: Sven Efftinge <sven.efftinge@typefox.io>