open workspace symbol: Don't omit entries without textual match
We are currently working on the workspace symbol support in clangd.
When the user types "ns1::func1", the workspace-symbol module wants to
generate an entry that looks like that.

  func1 | ns1:: - foo.c

Where "func1" is the label and "ns1:: - foo.c" the description part the
the quick open item. Then, when the quick-open module processes it, it
tries to fuzzy match what the user typed literally ("ns1::func1") in the
label or description (depending on the options). When there's no match,
it just drops the entry. So in this case, the entry shown above won't
be shown, even though it's a valid match.

This patch adds an option to the quick open mechanism to avoid filtering
out the entries that don't have an highlight (whether the highlight
comes from fuzzy matching or the data provider).

Signed-off-by: Simon Marchi simon.marchi@ericsson.com