Create a single logger for the process module
Currently, a new child logger for the 'process' module is created every
time a process is spawn (e.g. when a task is ran).  This can be observed
by placing a breakpoint in the Logger constructor (in the backend) and
running a task.

This could be a valid idea: we could add more information about the task
(e.g. its pid) to the logger, which would be included in each message
going through this logger).  However, we don't do it, which makes me
thing that it was not intentional to create a new logger for every task.

To simplify things (especially to implement #1600, Setting different log
levels per logger), this patch changes it to have a single logger for
the process module.  It uses the same technique of named bindings than
what the task and terminal modules already use.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>