Make the task code use the task logger
In the task module (task-backend-module.ts), we create a specific logger
to use (I assume) throughout the task code.  It is created using a named
binding like so:

    bind(ILogger).toDynamicValue(ctx => {
        const logger = ctx.container.get<ILogger>(ILogger);
        return logger.child({ 'module': 'task' });
    }).inSingletonScope().whenTargetNamed("task");

In order to get this logger instance, it is necessary to use the
@named('task') tag when injecting an ILogger instance, which we don't do
anywhere.  This means that this logger instance is not used currently.

This patch adds @named('task') to all injections of ILogger instances
throughout the task code.

Also, only the backend creates the 'task' specialized logger.  This
means that the task code in the frontend falls back on the root logger,
even if we try inject an ILogger instance with @named('task').  This
patch adds a similar binding in the frontend.  To avoid duplication, a
common module file is added for bindings which are identical in both the
frontend and the backend.

To avoid an error about ambiguous bindings when injecting an ILogger
without @named, we need to add .whenTargetIsDefault to the binding of
the root logger.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>