Allow user to search .gitignored files easily using ctrl+P
The quick file open menu currently displays all files, including those
ignored by .gitignore.  This results in a lot of noise when searching
for files.  This patch reverts the default to _not_ showing the ignored
files by default, but also adds a way to toggle this behavior.  It is
done by using the keyboard shortcut associated to the file search while
the file search popup is open.

Refs #1563

Signed-off-by: Jacques Bouthillier <jacques.bouthillier@ericsson.com>
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>