Set the root workspace when opening a new workspace
Right now you can open a new workspace with the preserveWindow option.
However when you set this one to true, it now reloads the current window
after having set the root in the default workspace server, but without
having updated the private _root property of the workspace service.

The problem arises when you the reload takes place, the onStop() hook
gets called and the service sets the root again for the server, however
as its root wasn't updated when open() was called, it sets the old root
in its place. The fix makes sure that the service has the appropriate
root set up before only when opening inside the same window, so that the
onStop hook can save the currently opening workspace correctly.

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@ericsson.com>