Re-add tslint.json file to use when building
This patch restores the previous behavior, which was to not have
warnings in the build output.  The rationale is that they are useful to
have in your editor, where you can fix them, but they're not really
interesting in the build output, since you probably won't do something
about it when seeing a warning there.

It adds some comments to the files to explain what they are used for.
Even if they are JSON files, the tslint.json format states that comments
are allowed:

https://palantir.github.io/tslint/usage/configuration/
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>